+++ src/prov/model/__init__.py
@@ -45,7 +45,7 @@ from prov.model.records import (
     AgentRef as AgentRef,
     DatetimeOrStr as DatetimeOrStr,
     EntityRef as EntityRef,
-    GenrationRef as GenrationRef,
+    GenerationRef as GenerationRef,
     Literal as Literal,
     NameValuePair as NameValuePair,
     NSCollection as NSCollection,