+++ src/prov/model/records.py
@@ -91,7 +91,7 @@ OptionalID = QualifiedNameCandidate | None  # type: typing.TypeAlias
 EntityRef = Union["ProvEntity", QualifiedNameCandidate]  # type: typing.TypeAlias
 ActivityRef = Union["ProvActivity", QualifiedNameCandidate]  # type: typing.TypeAlias
 AgentRef = Union["ProvAgent", "ProvEntity", "ProvActivity", QualifiedNameCandidate]  # type: typing.TypeAlias
-GenrationRef = Union["ProvGeneration", QualifiedNameCandidate]  # type: typing.TypeAlias
+GenerationRef = Union["ProvGeneration", QualifiedNameCandidate]  # type: typing.TypeAlias
 UsageRef = Union["ProvUsage", QualifiedNameCandidate]  # type: typing.TypeAlias
 RecordAttributesArg = (
     dict[QualifiedNameCandidate, Any] | Iterable[tuple[QualifiedNameCandidate, Any]]
@@ -1079,7 +1079,7 @@ class ProvEntity(ProvElement):
         self,
         usedEntity: EntityRef,
         activity: ActivityRef | None = None,
-        generation: GenrationRef | None = None,
+        generation: GenerationRef | None = None,
         usage: UsageRef | None = None,
         attributes: RecordAttributesArg | None = None,
     ) -> ProvEntity:
@@ -1162,7 +1162,7 @@ class ProvEntity(ProvElement):
         self,
         usedEntity: EntityRef,
         activity: ActivityRef | None = None,
-        generation: GenrationRef | None = None,
+        generation: GenerationRef | None = None,
         usage: UsageRef | None = None,
         attributes: RecordAttributesArg | None = None,
     ) -> ProvEntity:
@@ -1191,7 +1191,7 @@ class ProvEntity(ProvElement):
         self,
         usedEntity: EntityRef,
         activity: ActivityRef | None = None,
-        generation: GenrationRef | None = None,
+        generation: GenerationRef | None = None,
         usage: UsageRef | None = None,
         attributes: RecordAttributesArg | None = None,
     ) -> ProvEntity:
@@ -1220,7 +1220,7 @@ class ProvEntity(ProvElement):
         self,
         usedEntity: EntityRef,
         activity: ActivityRef | None = None,
-        generation: GenrationRef | None = None,
+        generation: GenerationRef | None = None,
         usage: UsageRef | None = None,
         attributes: RecordAttributesArg | None = None,
     ) -> ProvEntity: